api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
numpy==1.26.4
pillow==11.0.0
opencv-python==4.10.0.84
mediapipe==0.10.18
httpx==0.27.2
